// src/store/post/post.types.ts
enum POST_TYPES {
  GET_ALL = 'POST/GET_ALL',
  SET_ALL = 'POST/SET_ALL',
  GET_ONE = 'POST/GET_ONE',
  PATCH = 'POST/PATCH',
  DELETE = 'POST/DELETE',
  CREATE = 'POST/CREATE',
  CLEAR_CURRENT = 'POST/CLEAR_CURRENT',
  SET_ONE = 'POST/SET_ONE',
  SET_NEW_POST = 'POST/SET_NEW_ONE',
  SET_LOADING = 'POST/LOADING',
  SET_LOADING_SUCCESS = 'POST/LOADING_SUCCESS',
  SET_LOADING_FAILURE = 'POST/LOADING_FAILURE',
  SET_COUNT = 'POST/SET_COUNT',
  SET_PAGE = 'POST/SET_PAGE',
  DELETE_TAG = 'POST/DELETE_TAG',
  SET_USER_POSTS = 'POST/SET_USER_POST',
  SET_FILTER = 'POST/SET_FILTER',
  SET_ORDER = 'POST/SET_ORDER',
  SET_TAGS_ID = 'POST/SET_TAGS_ID',
  DELETE_TAG_FROM_POSTS = 'POSTS/DELETE_TAG',
}

export default POST_TYPES;